extensions completed

+++ src/controllers/booksController.js
@@ -3,7 +3,7 @@ const db = require('../../db')
 const booksRepository = require('../repositories/booksRepository')
 
 const getAllBooks = async (req, res) => {
-  const { type, topic, author } = req.query
+  const { type, topic, author, page, perPage } = req.query
   let values = []
   let query = ''
 
@@ -19,8 +19,13 @@ const getAllBooks = async (req, res) => {
     query = 'author'
     values = [author]
   }
-  const books = await booksRepository.getAllBooks(values, query)
-  res.json({ books })
+  if (perPage < 10 || perPage > 50) {
+    return res.status(400).json({ error: `parameter invalid perPage: ${perPage} not valid. Accepted range is 10 - 50`})
+  }
+  const books = await booksRepository.getAllBooks(values, query, page, perPage)
+  const perPageRes = Number(perPage)
+  const pageRes = Number(page)
+  res.json({ books, per_page: perPageRes, page: pageRes })
 }
 
 const getBookByID = async (req, res) => {
@@ -49,10 +54,10 @@ const updateBook = async (req, res) => {
   const bookTitleFound = await booksRepository.getBookByTitle([title])
 
   if (bookTitleFound) {
-    res.status(409).json({ error: `A book with the title: ${title} already exists`})
+    return res.status(409).json({ error: `A book with the title: ${title} already exists`})
   }
   if (!bookIdFound) {
-    res.status(404).json({ error: `no book with id: ${id}`})
+    return res.status(404).json({ error: `no book with id: ${id}`})
   }
   const book = await booksRepository.updateBook(values)
   res.status(201).json({ book })
@@ -62,6 +67,10 @@ const updateBook = async (req, res) => {
 const deleteBook = async (req, res) => {
   const { id } = req.params
   const values = [id]
+  const bookIdFound = await booksRepository.getBookByID([id])
+  if (!bookIdFound) {
+    return res.status(404).json({ error: `no book with id: ${id}`})
+  }
   const book = await booksRepository.deleteBook(values)
   res.status(201).json({ book })
 }
